Audio not audible on reboot.
Problem: While BT is streaming media and the device is rebooted. The audio
is no longer audible post reboot. The issue is that once the queue is
filled with exact 5 packets, it checks if audio focus is present or not.
If not present, it doesn't start the decoder thread. This results in queue
to be not flushed by the decoder and thus doesn't output any audio.

Solution: Start the sink decoder if it's not been started previously and
queue size has crossed the threshold and audio focus is already
acquired.

Test: m seahawk-userdebug. Tested on device by reproducing the
      scenario and validating other media use-cases.
Bug: 285956295
Tag: #stability
Ignore-AOSP-First: This is a CP from AOSP.
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:65df387f8ad45c887be7039be63be92f68c0f505)
Merged-In: Iff7ea4a79a9ec3f99930145d750174912dfa2ec9
Change-Id: Iff7ea4a79a9ec3f99930145d750174912dfa2ec9