Don't remove an account we haven't created yet
This change adds a boolean check to make sure we don't try to remove an
AccountManager account that we haven't created. Without the boolean
check, a java.lang.SecurityException can be thrown.

We risk getting into this state in the real world if we try to connect
and subsequently disconnect a device before our AccountManagerService is
up and associated with our process. Simple quick adapter toggles where
we try to connect to an out of range or off PBAP Server device can cause
the issue to very rarely appear.

Bug: 310662485
Bug: 348462216
Flag: EXEMPT, simple boolean check with a provided test
Test: atest PbapClientConnectionHandlerTest.java
Merged-In: Ic78e63a89d27ab5e7adf3b8d5458f1b61e85f83b
Change-Id: Ic78e63a89d27ab5e7adf3b8d5458f1b61e85f83b