LeAudio: Fix BidirectionalPair<> template instantiation
Use the implicit template instatiation for the BidirectionalPair<> as
it is easier to maintain, extend and use with additional types as
template arguments (e.g. in the unit tests).

There is no need for the explicit instatiation as there are no type
specific template specializations.

Bug: 380042533
Test: m -j
Flag: Exempt; No logical change
Change-Id: Ic3cbc872af54b70f13bfd550f9d969e7722fc3a2