Add VFS browsing Cover Art support
Cover art was implemented for browsing, but only when remote requested
Now Playing browsing. This was due to memory usage required to store images for
VFS browsing as the media players send an URI instead. We need to implement it,
and gate it behind a System property to enable as wanted to pass certification
tests. The implementation is already done in the native side. Also added PNG
variant support.

Bug: 364831930
Bug: 374271695
Test: atest BluetoothInstrumentationTests
Flag: exempt, change is disabled by default
Change-Id: Ib970e3add951bc7e9581e1e4b38b67d215828bb4