Update parcelable interface implementation to match tests for CTS-on-GSI
Problem: This branch is used to release our CTS test package, and
contains changes meant to stablize tests against the newest AOSP code.
When CTS is built from here and ran against aosp/main, everything is
fine. However, this branch is _also_ used to create the GSI image used
to replace the system partition in CTS-on-GSI based testing. This means
the source code here, however old, will replace that used on the build
GSI is applied to. Because the source code is old here, this can cause
desyncs between what the tests expect to be testing against and what the
actual implementation is. In this particular case,
BluetoothHeadsetClient was updated to follow errorprone recommendations
for parcelables on aosp/main. This impacted CTS testing on aosp/main,
which required a fix to the test to land here to stabilize testing. This
had the downstream impact of causing CTS-on-GSI to then fail, as the
source code here used to build GSI was no longer in sync with what the
tests expect.

Solution: Fix the implementation under test here so GSI can now match
what the test expects.

Bug: 370622991
Flag: Exempt, test only change
Test: atest CtsBluetoothTestCases:BluetoothHeadsetClientTest
Change-Id: I3e2410081235801697e4f8ceeae25d88a4a7916f