Make SetBrowsedPlayer return current path and correct item count.
Until now, SetBrowsedPlayer was returning the root id and the number
of items present in root folder. It should instead return the current
browsing path as changed by down and up events, and contain the
corresponding number of files of the current folder.

Bug: 297843126
Flag: com.android.bluetooth.flags.browsing_refactor
Test: atest BluetoothInstrumentationTests
Change-Id: Ifa0811c854389e25c961af2821d91590dfc41609