Drop messages with handles we didn't ask for
Problem: The typical message download process has us request the
contents for a specific message handle. Its possible that the remote
device can send back a message object with a handle we didn't recognize.
When this happens, the metadata for that message doesn't exist in our
map, which was causing an NPE.

Solution: Check that the handle exists in our map and don't use a null
object.

Flag: EXEMPT, simple null check with added test coverage
Bug: 373632001
Test: atest com.android.bluetooth.mapclient
Change-Id: I5a2efe865ce2fdbcd428b9c5d1eaedcbee0ff1bd