LeAudioService: Remove reduntant code
Group manager will be notified about unbonded device from the
bluetooth.cc so there is no need to call if from here.

Also, it has been noted that in some cases Java code is notified about
BOND_NONE state even the native stack is not really removing Bonding. In
such a case, device would be removed from the group and could not
recover from that situation on the reconnection time.
To recover, device would have to be unbonded

Bug: 381940827
Test: atest LeAudioServiceTest
Flag: Exempt, trivial fix
Change-Id: Iee5396b0be92dd2c395a014e6e3681b8af8518fb