LeAudio: Adjust target latency for TMAP & GMAP
This is according to the latest BT specification erratas.
Additionally, the voiceback channel reconfiguration
logic was adjusted to allow reconfigrations for the GAME scenario.
This parameter is used only with the codec extensibility logic, for
preparing the requirements when querying the codec config provider for
audio set configuration. It is not used when codec extensibility is
disabled or when using software encoding.

Bug: 365617442
Bug: 383623629
Test: atest bluetooth_le_audio_test
Flag: EXEMPT; day-to-day fix, regressions covered by the unit tests
Change-Id: Ib40e4790d2d97a2ec3ec6a16fa87136ed379d932