Also check the peer's name when applying INTEROP_DELAY_AUTH
It is reported on floss that some peer device consistently starting the
auth process about the same time as we also do, thus the two clashes
and it ended up in a failure.

Currently we have an interop in place but it only checks the address of
the peer. Our target peer device has quite a wide range of addresses
which is shared with a lot of non-target devices. Therefore, let's just
check the name as well to determine when to use the interop.

Bug: 382191454
Test: m -j
Flag: EXEMPT, trivial interop addition

Change-Id: I81d57fd17c9ba1f5ba40f93c72367bc9ffe9c1ff