avatar: allow to override tests
Allow avatar tests to be overridden, especially when they are from
external repository.
test_extended_scan is overridden for 4 specific configurations as they
trigger known bug: b/272120114 and fail. Since it is not known when this
bug will be fixed we need to skip those test configurations to not block
presubmit and development.

Bug: 272120114
Flag: EXEMPT - test only
Test: atest avatar:'LeHostTestFiltered' -v
Change-Id: I8b1f82be8aa0b3db9225a9289921927af936176d