floss: interop JBL Flip 5/6 to disregard HFP
JBL Flip 5/6 does not support HFP by specs and so is it in practice, but
it advertises the profile and that could cause trouble for Floss media
policy.

This CL adds an interop to disregard the HFP profile when advertised by
these headsets.

Bug: 381009230
Tag: #floss
Test: mmm packages/modules/Bluetooth
Flag: EXEMPT, enum adding and Floss-only logic change

Change-Id: I9079b0a877a3070cd1cbe7db9df202495e72f594