bta: change avdt open timeout for PTS test
In the test A2DP/SNK/SYN/BV-01-C the PTS sends open command
in around ~10 seconds, which is way longer than the stack allows
and it has to be handled in other way. This CL adds sysprop that will
setup the timeout to a greater value for PTS test purpose only.

Bug: 384490929
Flag: EXEMPT - test only
Test: atest pts-bot:A2DP/SNK/SYN/BV-01-C -v
Change-Id: Id5c4b7dbe2a176facee5eb843ddeaf785ca5c4e9