Check for decoding start even if we hit the max buffer size
Problem: A2DP Sink could get into a state where we would lock up with
the inability to decode. This was happening because we had logic that
would check the size of the decoder buffer/queue and remove the front
element to trim it back down to size if we needed. The logic would then
return instead of continuing on to check if we can start decoding. If
we ever filled the queue before decoding started, we would get into a
feedback loop where we would hit max size, remove an element, return,
hit max size again, remove an element, return, etc. We would never check
to see if we can start decoding the data we're adding.

Solution: Don't return, and instead check to see if we have focus and
enough data to start decoding.

Flag: EXEMPT, trivial bug fix
Bug: 367403820
Test: m com.android.bt
Change-Id: I58351e939c70d73804b8c144b9204c54ca46bf94