Bass: Always execute suspend procedure if external broadcast is managed
If broadcast was suspended (STATUS_LOCAL_STREAM_REQUESTED) and resuming
(STATUS_LOCAL_STREAM_SUSPENDED) was not finished before new suspend
come, then new suspend is not executed as sync status is checked before
that. Because information about paused broadcast sinks
(mPausedBroadcastSinks) is cleared at the beginning of the resume, lead
to situation that new resume is not executed too.

With this CL, suspend procedure will always be executed if the primary
device has managed external broadcast (no matter of sync status), so
sinks will be cached again even if previous resume was not finished.
With that, next resume will be executed properly.

Bug: 378615357
Bug: 378661060
Test: atest BassClientServiceTest
Change-Id: Ief1875712474cbae879c9e8445724bdcc03dad30