A2dp: relax getCodecStatus PRIVILEGED to allow cdm
Since aosp/3160584 getCodecStatus properly enforce the documented
permission.
This is breaking 3p apps that depends on reflection and hidden calls to
this API, despite supposedly be forbidden.

To adapt to current usage and to prepare a future eventual promotion to
public API (don't quote me :D), this CL remove the systematic
BLUETOOTH_PRIVILEGED enforcement and allow app with
companionDeviceManager association to retrieve the list of codec status.

Flag: Exempt trivial permission change
Test: m com.android.bt
Bug: 390289718
Bug: 389871692
Change-Id: Ib8dcb41af0e77364fcec2356836aa3331064ec74