LeAudio: Streamline the metadata usage
Make use of the LeAudioLtvMap data type instead of passing
the metadata via the raw uint8_t vector. This ensures that the
data is parsed only once and not every time it is needed. This way
we can detect invalid metadata at an early stage, in one place
where it is parsed the first (and now - the only) time.

Bug: 388733112
Test: atest --host bluetooth-test-audio-hal-le-audio-software bluetooth-test-audio-hal-aidl-leaudio-utils bluetooth_le_audio_test bluetooth_le_audio_client_test  bluetooth_test_broadcaster bluetooth_test_broadcaster_state_machine bluetooth_le_audio_codec_manager_test net_test_btif net_test_btif_hh net_test_btif_stack
Flag: EXEMPT; mechanical refactor - no logical change
Change-Id: Id833299701959c3f3ef39891a019a743dc85ce23