Fix addressed player not equals to active player
When the active player changes on the target, it should
also set the addressed player then update the remote
device. As this was made too late in the setActivePlayer
method, it would sometimes be skipped resulting in
mActivePlayer ID and mAddressedPlayer ID not in sync.
When these IDs are not in sync, the "play" events from
the remote are directly sent to the addresed player,
resulting in the wrong player resuming.

Bug: 388654447
Bug: 352279986
Flag: com.android.bluetooth.flags.set_addressed_player
Test: atest BluetoothInstrumentationTests
Change-Id: I8d9cc46357648060978dcd4853fc71145ca2dc4d