leaudio: Fix updating metadata for the group.
Each group keeps its metadata so whenever set member is trying to attach
to the stream, it is verfing if its available contexts matches group
metadata.

This CL fixes updates of this metadata which could be 0x0000
when given group was configured before the  metadata were known.

In the real life it can happen, when Le Headset is connecting during the
call and client.cc is trying to configure it as soon as possible which
happens in GroupSetActive() function. At that time, if there were not
previous LeAudio Headset connected, Bluetooth is not aware of the
metadata, however it knows that bidirectional stream will be needed for
the phone call and configuration is happening.

Without this CL, when Audio HAL updated Bluetooth with metadata, those
were not set into the group, which later leads to situation that second
attached set member was no connected to the stream due to missing
metadata.

Bug: 392012679
Flag: Exempt, day-to-day fix. Regression tested with unit test.
Test: atest bluetooth_le_audio_test

Change-Id: Ib29d7d70ece6fee8c5c67150b5211b00002bb84c