InCallService: stop self enforcing permission
Enforcing permission here is a non-sense as we are making the
call to the method and then checking the permission while never leaving
the bluetooth boundary.

More over, if Bluetooth don't have the permission, this would crash the
Bluetooth app entirely as we are throwing in our own process

This is one of the reason while in https://r.android.com/3468588, the
associated test cannot be run as a normal unit test and need the
instrumentation.

Bug: 191783947
Test: m .
Flag: Exempt refactor no-op
Change-Id: I53daa16cab5982bf00a3f33f043873d7321f509a