InCallService: clccResponse stays within Bt
The BluetoothInCallService service is living within the Bluetooth app
process.
Calling the Bluetooth Framework like if it was a 3p app doesn't make
sense and this is preventing efficient testing with un-necessary
abstraction.
This is also exposing some hidden API that shouldn't be callable.

This change replace the call to the bluetooth framework by a direct call
to the internal profile.

Bug: 380356012
Test: atest BluetoothInstrumentationTests
Flag: Exempt no-op refactor | this is simply shortening the calls to the
      service
Change-Id: I7abc6d4b1e647ab35170d709e713b393b219ee03