Bass: Add only one sink to addSource while BassStateReady
handleBassStateReady gives information about readiness of BASS for one
sink. Because of that, there is need to addSource only for this one, not
for whole group. Second device has its own object in
mPendingSourcesToAdd.

enqueueSourceGroupOp will not be set in addSource then, which is wrong,
but it should be fixed in another CL globally as this is broken in whole
handleBassStateReady and resumeReceiversSourceSynchronization.

Bug: 397164105
Flag: EXEMPT, day-to-day bugfix covered by unit tests
Test: atest BassClientServiceTest
Change-Id: I9a76974acd44ebfb1c97d96b0c632420735997a7