pbapclient: Handler fixes
PbapClientStateMachineOld is implicitly creating an handler for each
devices, (and a second explictlyin the onEnter of connecting).
This CL does one no-op change to explicitly defined the handler to make
it obvious.
This CL does the nit behavior changes to wait on the handler before
cleaning up, as this is a race condition and is currently failing our
test infra since mock are reset too quickly and the thread crash
This is due to the test not controlling its execution

Bug: 398953257
Test: atest BluetoothJavaUnitTests
Flag: TEST_ONLY
Change-Id: I3902d6e2e2d9bf009f17fe21b1a029e4dd9b6277