HID Host: Reset conn state if failed to disconnect
It's possible for the ctrl_cid to be populated even though no l2cap
connect has been issued. When the baseband connection is disconnected
under this situation, we still send L2CAP DisconnectReq but it will
fail and we won't get a response.

Call the response callback in this case.

Bug: 344778226
Bug: 387937345
Test: m -j
Test: Connect a classic HID device, then toggle the "Input" selection
      on the device details page. Verify device is still connectable.
Flag: com.android.bluetooth.flags.disconnect_hid_channels_serially

Change-Id: I97ca0d6f6fbf6a6ed773c3d68ea037352758e839