leaudio: Do not clear codec preference when invalidating context map cache
Codec preference is coming from the Java layer, and are kept in preferred_config_.
When configuriation is triggered for the group, there is context_to_preferred_configuration_cache_map_ filled out with the configurations. Similar to what happens with context_to_configuration_cache_map_ when there is no preference provided.

When this cache is invalidated with InvalidateCachedConfigurations(), we should just clear two cache_maps_ instead of the Java preference as those will be used later during configuration.

We want this preference to stay valid as long as Bluetooth is enabled.

Bug: 331775328
Test: atest bluetooth_le_audio_test bluetooth_le_audio_client_test
Flag: Exempt, day-to-day fix, regression covered with unit test
Change-Id: Id5791d33402813f07346af0ee82e1a75e36c8863