Re-order SnoopLoggerTracing flush and bundling.
Flushing after bundling includes the first packet after the gap. For
exmaple, if a CMD event is recorded at t=10s and another at t=20s, the
previous implementation would bundle both. This creates an event with
with a duration of 10 seconds which is undesirable. This change first
flushes the old event, and then starts a new bundle for the new event.

Bug: 365778358
Test: m sync
Test: build, flash, and record trace
Change-Id: Ia5419dd797fe5103c3984fcc6e30751489adc185